refactor(server): migrate /graph/v3/query endpoint to v3 layer
